Add with_fee test, split out helper utils and clean up tests
